Moved some functionality out of the Connection class to make it smaller.
The idea is to separate the protocol parsing from the socket and request
handling. We should probably rename some classes to make their purpose
more clear. The reason for the separation is that we want to be able to
integrate with gevent, without having to re-implement anything related to
the protocol parsing.